Restructure project to be more organized/logical.
The project has been restructured around different "modules" each with their own actions, reducer, components, etc. README will explain structure a bit better.
