--- src/js/components/connectedHome.js
@@ -1,15 +1,0 @@
-"use strict";
-
-import {connect} from 'react-redux';
-import Home from './home';
-
-const ConnectedHome = connect(
-    // state => ({
-    //     //...
-    // }),
-    // dispatch => ({
-
-    // })
-)(Home);
-
-export default ConnectedHome;
