--- src/js/components/views/loginView/connectedLoginView.js
@@ -1,19 +1,0 @@
-"use strict";
-
-import { connect } from 'react-redux';
-import LoginView from './loginView';
-
-import { requestLogin } from '../../../actions/authentication';
-
-const ConnectedLoginView = connect(
-    state => ({
-        isAuthenticated: state.authentication.isAuthenticated,
-        isAuthenticating: state.authentication.isFetching
-    }),
-    dispatch => ({
-        attemptLogin: authCode => dispatch(requestLogin(authCode))
-    })
-)(LoginView);
-
-
-export default ConnectedLoginView;
