--- src/js/components/views/submissionView/connectedSubmissionView.js
@@ -1,11 +1,0 @@
-"use strict";
-
-import { connect } from 'react-redux';
-import SubmissionView from './submissionView';
-
-const ConnectedSubmissionView = connect(
-    () => {},
-    () => {}
-)(SubmissionView);
-
-export default ConnectedSubmissionView;
