+++ src/js/lib/components/authentication/connectedAuthentication.js
