+++ src/js/lib/components/editableLabel.jsx
