+++ src/js/lib/components/loadingAnimations/defaultLoadingAnimation.jsx
