+++ src/js/modules/authentication/actionTypes.js
@@ -1,0 +1,9 @@
+"use strict";
+
+export const TOKEN_LOADED = "authentication/TOKEN_LOADED";
+export const NO_TOKEN_LOADED = "authentication/NO_TOKEN_LOADED";
+
+export const REQUEST_LOGIN = "authentication/REQUEST_LOGIN";
+export const SUCCESS_LOGIN = "authentication/SUCCESS_LOGIN";
+export const FAIL_LOGIN = "authentication/FAIL_LOGIN";
+export const LOGOUT = "authentication/LOGOUT";
