+++ src/js/modules/authentication/constants.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const name = "authentication";
