+++ src/js/modules/authentication/index.js
@@ -1,0 +1,17 @@
+"use strict";
+
+import reducer from './reducer';
+import * as constants from './constants';
+import LoginView from './components/loginView';
+import * as actionTypes from './actionTypes';
+import * as actions from './actions';
+
+export default {
+    components: {
+        LoginView
+    },
+    actions,
+    actionTypes,
+    constants,
+    reducer
+};
