+++ src/js/modules/home/actionTypes.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const ACTION_NAME = "home/ACTION_NAME";
