+++ src/js/modules/home/components/homeView/index.js
@@ -1,0 +1,15 @@
+"use strict";
+
+import {connect} from 'react-redux';
+import HomeView from './homeView';
+
+const ConnectedHomeView = connect(
+    // state => ({
+    //     //...
+    // }),
+    // dispatch => ({
+
+    // })
+)(HomeView);
+
+export default ConnectedHomeView;
