+++ src/js/modules/home/constants.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const name = "home";
