+++ src/js/modules/home/index.js
@@ -1,0 +1,13 @@
+"use strict";
+
+import reducer from './reducer';
+import * as constants from './constants';
+import HomeView from './components/homeView';
+
+export default {
+    components: {
+        HomeView
+    },
+    constants,
+    reducer
+};
