+++ src/js/modules/navigation/actionTypes.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const ACTION_NAME = "navigation/ACTION_NAME";
