+++ src/js/modules/navigation/actions.js
@@ -1,0 +1,9 @@
+"use strict";
+
+import * as actions from './actionTypes';
+
+export function actionName() {
+    return {
+        type: actions.ACTION_NAME
+    };
+}
