+++ src/js/modules/navigation/index.js
@@ -1,0 +1,13 @@
+"use strict";
+
+import reducer from './reducer';
+import * as constants from './constants';
+import NavigationFrame from './components/navigationFrame';
+
+export default {
+    components: {
+        NavigationFrame
+    },
+    constants,
+    reducer
+};
