+++ src/js/modules/submission/actionTypes.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const ACTION_NAME = "submission/ACTION_NAME";
