+++ src/js/modules/submission/components/submissionView/index.js
@@ -1,0 +1,11 @@
+"use strict";
+
+import { connect } from 'react-redux';
+import SubmissionView from './submissionView';
+
+const ConnectedSubmissionView = connect(
+    () => ({}),
+    () => ({})
+)(SubmissionView);
+
+export default ConnectedSubmissionView;
