+++ src/js/modules/userInfo/actionTypes.js
@@ -1,0 +1,9 @@
+"use strict";
+
+export const REQUEST_USER_INFO = "userInfo/REQUEST_USER_INFO";
+export const REFRESH_USER_INFO = "userInfo/REFRESH_USER_INFO";
+export const SUCCESS_USER_INFO = "userInfo/SUCCESS_USER_INFO";
+export const FAIL_USER_INFO = "userInfo/FAIL_USER_INFO";
+export const REQUEST_UPDATE_USER_INFO = "userInfo/REQUEST_UPDATE_USER_INFO";
+export const SUCCESS_UPDATE_USER_INFO = "userInfo/SUCCESS_UPDATE_USER_INFO";
+export const FAIL_UPDATE_USER_INFO = "userInfo/FAIL_UPDATE_USER_INFO";
