+++ src/js/modules/userInfo/constants.js
@@ -1,0 +1,3 @@
+"use strict";
+
+export const name = "userInfo";
